+++ pulsar/internal/buffer.go
@@ -17,7 +17,7 @@
 // under the License.
 //
 
-package impl
+package internal
 
 import (
 	"encoding/binary"
