+++ pulsar/internal/closable.go
@@ -17,12 +17,8 @@
 // under the License.
 //
 
-package impl
+package internal
 
-import "hash/crc32"
-
-var crc32cTable = crc32.MakeTable(crc32.Castagnoli)
-
-func Crc32cCheckSum(data []byte) uint32 {
-	return crc32.Checksum(data, crc32cTable)
+type Closable interface {
+	Close() error
 }
