+++ pulsar/internal/default_router.go
@@ -17,7 +17,7 @@
 // under the License.
 //
 
-package impl
+package internal
 
 import (
 	"math/rand"
