+++ pulsar/internal/default_router_test.go
@@ -17,7 +17,7 @@
 // under the License.
 //
 
-package impl
+package internal
 
 import (
 	"github.com/stretchr/testify/assert"
