+++ pulsar/internal/topic_name.go
@@ -17,7 +17,7 @@
 // under the License.
 //
 
-package impl
+package internal
 
 import (
 	"errors"
