+++ pulsar/producer_test.go
@@ -23,7 +23,7 @@ import (
 	"context"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
-	"pulsar-client-go/pulsar/impl/util"
+	"pulsar-client-go/pulsar/internal/util"
 	"sync"
 	"testing"
 	"time"
